Tests/LibWeb/Text/input/abspos-relayout-scroller.html

<!DOCTYPE html>
<script src="include.js"></script>
<style>
    body { margin: 0; }
    #parent { position: relative; width: 220px; height: 160px; }
    #boundary {
        position: absolute;
        left: 20px;
        top: 30px;
        width: 100px;
        height: 100px;
        overflow: auto;
    }
    #content {
        width: 50px;
        height: 500px;
        box-sizing: border-box;
        padding-top: 180px;
    }
    #probe { height: 20px; }
</style>
<div id="parent">
    <div id="boundary">
        <div id="content">
            <div id="probe">probe</div>
        </div>
    </div>
</div>
<script>
    asyncTest(async (done) => {
        document.body.offsetHeight;

        const scroller = document.getElementById("boundary");
        scroller.scrollTop = 150;

        const partialBefore = internals.partialLayoutCount();
        const fullBefore = internals.fullLayoutCount();
        document.getElementById("content").style.height = "520px";
        document.body.offsetHeight;
        const partialDelta = internals.partialLayoutCount() - partialBefore;
        const fullDelta = internals.fullLayoutCount() - fullBefore;

        const hit = document.elementFromPoint(30, 65);
        println(`scrollTop=${scroller.scrollTop}`);
        println(`hit=${hit ? (hit.id || hit.tagName) : "null"}`);
        println(`partial-layouts=${partialDelta} full-layouts=${fullDelta}`);
        println(scroller.scrollTop === 150 && hit && hit.id === "probe"
            && partialDelta === 1 && fullDelta === 0 ? "PASS" : "FAIL");
        done();
    });
</script>
