Tests/LibWeb/Text/expected/abspos-relayout-scroller.txt

scrollTop=150
hit=probe
partial-layouts=0 full-layouts=1
FAIL
